Fix Content-Length format specifier in https_server

+++ chap10/https_server.c
@@ -249,7 +249,7 @@ void serve_resource(struct client_info *client, const char *path) {
     sprintf(buffer, "Connection: close\r\n");
     SSL_write(client->ssl, buffer, strlen(buffer));
 
-    sprintf(buffer, "Content-Length: %u\r\n", cl);
+    sprintf(buffer, "Content-Length: %lu\r\n", (unsigned long)cl);
     SSL_write(client->ssl, buffer, strlen(buffer));
 
     sprintf(buffer, "Content-Type: %s\r\n", ct);
